+++ src/@batch-flask/core/data/entity-getter/index.ts
@@ -1,0 +1,1 @@
+export * from "./entity-getter";
