+++ src/@batch-flask/core/data/entity-view/index.ts
@@ -1,0 +1,1 @@
+export * from "./entity-view";
